requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.2
cassandra-driver==3.28.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
numpy==1.24.3
sentence-transformers==2.2.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1 